requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
dj-database-url==3.0.1
Django==5.2.5
django-allauth==65.11.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mysqlclient==2.2.7
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.9.0
